agents/langgraph-agent/requirements.txt
langchain
langchain-community
langchain-google-genai
langgraph
chromadb
python-dotenv